Issue #922 - Implements methods Connection.getBytes[In|Out]().
Using int rather than long as return value for generated bytes.

+++ jetty-http2/http2-common/src/main/java/org/eclipse/jetty/http2/HTTP2Session.java
@@ -1108,7 +1108,7 @@ public abstract class HTTP2Session extends ContainerLifeCycle implements ISessio
 
     private class ControlEntry extends HTTP2Flusher.Entry
     {
-        private long bytes;
+        private int bytes;
 
         private ControlEntry(Frame frame, IStream stream, Callback callback)
         {
@@ -1224,7 +1224,7 @@ public abstract class HTTP2Session extends ContainerLifeCycle implements ISessio
     private class DataEntry extends HTTP2Flusher.Entry
     {
         private int length;
-        private long bytes;
+        private int bytes;
 
         private DataEntry(DataFrame frame, IStream stream, Callback callback)
         {
