+++ jetty-http2/http2-common/src/main/java/org/eclipse/jetty/http2/generator/PingGenerator.java
@@ -35,13 +35,13 @@ public class PingGenerator extends FrameGenerator
     }
 
     @Override
-    public long generate(ByteBufferPool.Lease lease, Frame frame)
+    public int generate(ByteBufferPool.Lease lease, Frame frame)
     {
         PingFrame pingFrame = (PingFrame)frame;
         return generatePing(lease, pingFrame.getPayload(), pingFrame.isReply());
     }
 
-    public long generatePing(ByteBufferPool.Lease lease, byte[] payload, boolean reply)
+    public int generatePing(ByteBufferPool.Lease lease, byte[] payload, boolean reply)
     {
         if (payload.length != PingFrame.PING_LENGTH)
             throw new IllegalArgumentException("Invalid payload length: " + payload.length);
